Update README.md
Objective: generate a square wave based on user input.

The user will be prompted to enter the frequency and duty cycle of the square signal on the LCD screen. 

The user will use a 4×4 keypad to enter the frequency and duty cycle.

The user input will be echoed (displayed) on the screen.

Invalid keys will be filtered (ie if user enters 123A then the screen will only display 123 and the A is disregarded, and the user will be allowed to continue entering values).

Don’t allow the user to enter an out of range number (i.e. Input can’t exceed the max for frequency and duty cycle. Max for duty cycle is 100. Max for frequency is be determined by you). 

Convert the user input to a number “on the fly” (ie update the number as the user is inputting digits and don’t store the input into a string array first). 

Extra credit: allow the user to “backspace” (delete input 1 character a time).

The LCD 4-bit mode is preferred, but 8-bit mode is OK. 

The program should repeat until user decides to quit.

You should build your project in Simulation IDE and use Microchip Studio for programming.

Use the oscilloscope in Simulation IDE to confirm the wave generation.